Drop the use of 'union ALIGNMENT payload[0]'
MSVC cannot handle zero sized arrays in the middle of a structure.
Also, the use of forced alignment after 3 pointers and a size_t in
a struct appears very unnecessary, as, unless one is using a broken
compiler, whatever follows is always aligned to 64-bit, regardless
of whether the platform is 32 or 64-bit.

Therefore, remove 'union ALIGNMENT' altogether and reference
((struct CACHED_INODE*)<ptr>)->inum instead of <ptr>->payload to
accomplish the thing.