Fix MSVC incompatibility with test_and_[set|clear]_bit macros
The use of macros for test_and_set_bit() and test_and_clear_bit()
produces syntax errors for MSVC due incompatible GCC-isms.
Switch to using static inline calls instead, which should be more
universally supported.

Note that, because we need to pass variables by reference (so that
the actual variable is modified instead of just a local copy) we
still define some macros to call into the new inline functions.