Add UEFI CompilerIntrinsicsLib for EDK2/MSVC x86 compilation
EDK2 compilation using MSVC is lacking some compiler intrinsics
that it is not possible to simply redefine in the driver source
due to restrictions placed by Microsoft on which intrinsics can
be (re)defined in regular code.

As such, we have to use a formal UEFI library to provide the calls
we need, in the same manner as the EDK2 also provides a similar
library for ARM/ARM64 architectures.